use PHONES_ENABLED env var to better separate phones app
