+++ .circleci/config.yml
@@ -173,6 +173,7 @@ jobs:
             docker run \
               --entrypoint "/bin/bash" \
               --volumes-from workspace-test-results \
+              -e PHONES_ENABLED=$PHONES_ENABLED \
               -e TWILIO_ACCOUNT_SID=$TWILIO_ACCOUNT_SID \
               -e TWILIO_AUTH_TOKEN=$TWILIO_AUTH_TOKEN \
               fx-private-relay \
