+++ .env-dist
@@ -27,6 +27,7 @@ RELAY_FROM_ADDRESS="relay@127.0.0.1:8000"
 NEW_RELAY_FROM_ADDRESS="new_from@127.0.0.1:8000"
 SITE_ORIGIN="http://127.0.0.1:8000"
 MAX_NUM_FREE_ALIASES=5
+PHONES_ENABLED=False
 TWILIO_ACCOUNT_SID=
 TWILIO_AUTH_TOKEN=
 TWILIO_SMS_APPLICATION_SID=
