+++ api/serializers/__init__.py
@@ -1,9 +1,8 @@
+from django.contrib.auth.models import User
+
 from rest_framework import serializers, exceptions
 
 from emails.models import Profile, DomainAddress, RelayAddress
-from django.contrib.auth.models import User
-
-from phones.models import RealPhone, RelayNumber
 
 
 class PremiumValidatorsMixin:
@@ -141,37 +140,3 @@ class UserSerializer(serializers.ModelSerializer):
         model = User
         fields = ["email"]
         read_only_fields = ["email"]
-
-
-class RealPhoneSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = RealPhone
-        fields = [
-            "id",
-            "number",
-            "verification_code",
-            "verification_sent_date",
-            "verified",
-            "verified_date",
-        ]
-        read_only_fields = [
-            "id",
-            "verification_sent_date",
-            "verified",
-            "verified_date",
-        ]
-        extra_kwargs = {
-            "verification_code": {"write_only": True},
-        }
-
-
-class RelayNumberSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = RelayNumber
-        fields = [
-            "number",
-            "location",
-        ]
-        read_only_fields = [
-            "location",
-        ]
