+++ api/urls.py
@@ -3,7 +3,15 @@ from django.urls import include, path, register_converter
 
 from rest_framework import routers
 
-from . import views
+from .views import (
+    DomainAddressViewSet,
+    RelayAddressViewSet,
+    ProfileViewSet,
+    UserViewSet,
+    premium_countries,
+    runtime_data,
+    schema_view,
+)
 
 
 class SwaggerFormatConverter:
@@ -20,29 +28,37 @@ register_converter(SwaggerFormatConverter, "swagger_format")
 
 
 api_router = routers.DefaultRouter()
-api_router.register(r"domainaddresses", views.DomainAddressViewSet, "domainaddress")
-api_router.register(r"relayaddresses", views.RelayAddressViewSet, "relayaddress")
-api_router.register(r"profiles", views.ProfileViewSet, "profiles")
-api_router.register(r"users", views.UserViewSet, "user")
-if "phones.apps.PhonesConfig" in settings.INSTALLED_APPS:
-    api_router.register(r"realphone", views.RealPhoneViewSet, "real_phone")
-    api_router.register(r"relaynumber", views.RelayNumberViewSet, "relay_number")
+api_router.register(r"domainaddresses", DomainAddressViewSet, "domainaddress")
+api_router.register(r"relayaddresses", RelayAddressViewSet, "relayaddress")
+api_router.register(r"profiles", ProfileViewSet, "profiles")
+api_router.register(r"users", UserViewSet, "user")
 
 
 urlpatterns = [
-    path("v1/premium_countries", views.premium_countries, name="premium_countries"),
-    path("v1/runtime_data", views.runtime_data, name="runtime_data"),
-    path("v1/inbound_sms", views.inbound_sms, name="inbound_sms"),
-    path("v1/vCard/<lookup_key>", views.vCard, name="vCard"),
+    path("v1/premium_countries", premium_countries, name="premium_countries"),
+    path("v1/runtime_data", runtime_data, name="runtime_data"),
     path(
         "v1/swagger<swagger_format:format>",
-        views.schema_view.without_ui(cache_timeout=0),
+        schema_view.without_ui(cache_timeout=0),
         name="schema-json",
     ),
     path(
         "v1/docs/",
-        views.schema_view.with_ui("swagger", cache_timeout=0),
+        schema_view.with_ui("swagger", cache_timeout=0),
         name="schema-swagger-ui",
     ),
+]
+
+if settings.PHONES_ENABLED:
+    from .views.phones import RealPhoneViewSet, RelayNumberViewSet, inbound_sms, vCard
+
+    api_router.register(r"realphone", RealPhoneViewSet, "real_phone")
+    api_router.register(r"relaynumber", RelayNumberViewSet, "relay_number")
+    urlpatterns += [
+        path("v1/inbound_sms", inbound_sms, name="inbound_sms"),
+        path("v1/vCard/<lookup_key>", vCard, name="vCard"),
+    ]
+
+urlpatterns += [
     path("v1/", include(api_router.urls)),
 ]
