+++ privaterelay/settings.py
@@ -173,6 +173,7 @@ RECRUITMENT_BANNER_TEXT = config("RECRUITMENT_BANNER_TEXT", None)
 RECRUITMENT_EMAIL_BANNER_TEXT = config("RECRUITMENT_EMAIL_BANNER_TEXT", None)
 RECRUITMENT_EMAIL_BANNER_LINK = config("RECRUITMENT_EMAIL_BANNER_LINK", None)
 
+PHONES_ENABLED = config("PHONES_ENABLED", False, cast=bool)
 TWILIO_ACCOUNT_SID = config("TWILIO_ACCOUNT_SID", None)
 TWILIO_AUTH_TOKEN = config("TWILIO_AUTH_TOKEN", None)
 TWILIO_MAIN_NUMBER = config("TWILIO_MAIN_NUMBER", None)
@@ -229,7 +230,7 @@ if AWS_SES_CONFIGSET and AWS_SNS_TOPIC:
         "emails.apps.EmailsConfig",
     ]
 
-if TWILIO_ACCOUNT_SID and TWILIO_AUTH_TOKEN:
+if PHONES_ENABLED and TWILIO_ACCOUNT_SID and TWILIO_AUTH_TOKEN:
     INSTALLED_APPS += [
         "phones.apps.PhonesConfig",
     ]
